Availability of 0.1.0-alpha.2

+++ crates/sklears-impute/src/categorical/hot_deck.rs
@@ -134,7 +134,7 @@ impl Fit<ArrayView2<'_, Float>, ()> for HotDeckImputer<Untrained> {
 
     #[allow(non_snake_case)]
     fn fit(self, X: &ArrayView2<'_, Float>, _y: &()) -> SklResult<Self::Fitted> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         // If no categorical features specified, assume all are categorical
@@ -217,7 +217,7 @@ impl Fit<ArrayView2<'_, Float>, ()> for HotDeckImputer<Untrained> {
 impl Transform<ArrayView2<'_, Float>, Array2<Float>> for HotDeckImputer<HotDeckImputerTrained> {
     #[allow(non_snake_case)]
     fn transform(&self, X: &ArrayView2<'_, Float>) -> SklResult<Array2<Float>> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         if n_features != self.state.n_features_in_ {
