+++ crates/sklears-impute/src/kernel/gaussian_process.rs
@@ -375,7 +375,7 @@ impl Fit<ArrayView2<'_, Float>, ()> for GaussianProcessImputer<Untrained> {
 
     #[allow(non_snake_case)]
     fn fit(self, X: &ArrayView2<'_, Float>, _y: &()) -> SklResult<Self::Fitted> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         // Collect complete cases for training
@@ -498,7 +498,7 @@ impl Transform<ArrayView2<'_, Float>, Array2<Float>>
 {
     #[allow(non_snake_case)]
     fn transform(&self, X: &ArrayView2<'_, Float>) -> SklResult<Array2<Float>> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         if n_features != self.state.n_features_in_ {
@@ -566,7 +566,7 @@ impl GaussianProcessImputer<GaussianProcessImputerTrained> {
         &self,
         X: &ArrayView2<'_, Float>,
     ) -> SklResult<Vec<Vec<GPPredictionResult>>> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         if n_features != self.state.n_features_in_ {
