+++ crates/sklears-impute/src/kernel/svr.rs
@@ -198,7 +198,7 @@ impl Fit<ArrayView2<'_, Float>, ()> for SVRImputer<Untrained> {
 
     #[allow(non_snake_case)]
     fn fit(self, X: &ArrayView2<'_, Float>, _y: &()) -> SklResult<Self::Fitted> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         // Collect complete cases for training
@@ -285,7 +285,7 @@ impl Fit<ArrayView2<'_, Float>, ()> for SVRImputer<Untrained> {
 impl Transform<ArrayView2<'_, Float>, Array2<Float>> for SVRImputer<SVRImputerTrained> {
     #[allow(non_snake_case)]
     fn transform(&self, X: &ArrayView2<'_, Float>) -> SklResult<Array2<Float>> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         if n_features != self.state.n_features_in_ {
