+++ crates/sklears-impute/src/matrix_factorization.rs
@@ -139,7 +139,7 @@ impl Fit<ArrayView2<'_, Float>, ()> for MatrixFactorizationImputer<Untrained> {
     type Fitted = MatrixFactorizationImputer<MatrixFactorizationImputerTrained>;
 
     fn fit(self, X: &ArrayView2<'_, Float>, _y: &()) -> SklResult<Self::Fitted> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         if self.n_components > n_features.min(n_samples) {
@@ -323,7 +323,7 @@ impl Transform<ArrayView2<'_, Float>, Array2<Float>>
     for MatrixFactorizationImputer<MatrixFactorizationImputerTrained>
 {
     fn transform(&self, X: &ArrayView2<'_, Float>) -> SklResult<Array2<Float>> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         if n_features != self.state.n_features_in_ {
