+++ crates/sklears-impute/src/tree_methods.rs
@@ -120,7 +120,7 @@ impl Fit<ArrayView2<'_, Float>, ()> for DecisionTreeImputer<Untrained> {
     type Fitted = DecisionTreeImputer<DecisionTreeImputerTrained>;
 
     fn fit(self, X: &ArrayView2<'_, Float>, _y: &()) -> SklResult<Self::Fitted> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (_, n_features) = X.dim();
 
         // Calculate feature means for fallback
@@ -209,7 +209,7 @@ impl Transform<ArrayView2<'_, Float>, Array2<Float>>
     for DecisionTreeImputer<DecisionTreeImputerTrained>
 {
     fn transform(&self, X: &ArrayView2<'_, Float>) -> SklResult<Array2<Float>> {
-        let X = X.mapv(|x| x as f64);
+        let X = X.mapv(|x| x);
         let (n_samples, n_features) = X.dim();
 
         if n_features != self.state.n_features_in_ {
